requirements.txt
torch==1.7.1
gdown==3.13.0
tqdm==4.60.0
numpy==1.19.5
opencv_python==4.5.1.48
torchvision==0.8.2
seaborn==0.11.1
matplotlib==3.3.4
openpifpaf==0.13.0
scikit_image==0.17.2
Pillow==8.4.0
scikit_learn
scikit-image